# backend/rest/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API para emissão de Badges",
    "description": "\n## Token de acesso\nEnvia email e senha via Post como o comando de exemplo:\n```bash\ncurl -X POST 'https://{host:port}/user/login' -d \"email=EMAIL&password=SENHA\"\n```\n\nResposta:\n```javascript\n{\n    \"access_token\": \"TOKENDEACESSO\",\n    \"token_type\": \"Bearer\",\n    \"expires_in\": 86400,\n    \"refresh_token\": \"TOKENDEREFRESH\",\n}\n```\n\n## Expiração de Token\nToken de expiração para renovar  por examplo:\n\n```bash\ncurl -X POST 'https://{host:port}/user/refresh' -d \"refreshToken=refresh_token```\n\n"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "any description..."
    }
  },
  "security": [{"AuthorizationHeader": []}],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/assertions/file/{id}": {
      "get": {
        "description": "Busca um arquivo de badge",
        "tags":["Assertions"],
        "security": [ { "bearerAuth": [] } ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Negado"
          }
        }
      }
    },
...
    "/files/{file}": {
      "get": {
        "description": "Busca um arquivo no servidor de badges",
        "security": [ { "bearerAuth": [] } ],
        "tags":["Files"],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "file",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Negado"
          }
        }
      }
    }
  },
...
}